+++ CHANGELOG.md
@@ -5,7 +5,74 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.10] - 2026-06-17
+
+### Fixed
+- Fix ai-boilerplate issues (ticket-0667fda4)
+- Fix unused-imports issues (ticket-2d23d774)
+- Fix unused-imports issues (ticket-86a3e5b9)
+- Fix magic-numbers issues (ticket-42824761)
+- Fix ai-boilerplate issues (ticket-536cbb66)
+- Fix string-concat issues (ticket-197e148c)
+- Fix unused-imports issues (ticket-89afe2cd)
+- Fix string-concat issues (ticket-147926e7)
+- Fix unused-imports issues (ticket-e0d45a9e)
+- Fix unused-imports issues (ticket-d64347fc)
+- Fix string-concat issues (ticket-1a6d1f16)
+- Fix unused-imports issues (ticket-b922a8ca)
+- Fix unused-imports issues (ticket-ea8cb18d)
+- Fix string-concat issues (ticket-28b42651)
+- Fix unused-imports issues (ticket-04d93ec8)
+- Fix string-concat issues (ticket-679c43e4)
+- Fix unused-imports issues (ticket-11918242)
+- Fix string-concat issues (ticket-34b4ee88)
+- Fix unused-imports issues (ticket-f0288fc8)
+- Fix unused-imports issues (ticket-3b47be76)
+- Fix smart-return-type issues (ticket-adfadfb0)
+- Fix string-concat issues (ticket-2de679dd)
+- Fix magic-numbers issues (ticket-d19c4d7f)
+- Fix string-concat issues (ticket-d35f9c4d)
+- Fix unused-imports issues (ticket-e8ab6580)
+- Fix smart-return-type issues (ticket-06bbd30d)
+- Fix string-concat issues (ticket-b5d2908d)
+- Fix unused-imports issues (ticket-23f58514)
+- Fix magic-numbers issues (ticket-c63b1916)
+- Fix smart-return-type issues (ticket-f5104d30)
+- Fix smart-return-type issues (ticket-a9aca8a7)
+- Fix unused-imports issues (ticket-80f3da79)
+- Fix smart-return-type issues (ticket-7fecfff1)
+
 ## [Unreleased]
+
+## [0.1.2] - 2026-06-17
+
+### Docs
+- Update CHANGELOG.md
+- Update README.md
+- Update SUMD.md
+- Update SUMR.md
+- Update TODO.md
+- Update project/README.md
+- Update project/context.md
+
+### Test
+- Update testql-scenarios/generated-api-smoke.testql.toon.yaml
+- Update testql-scenarios/generated-cli-tests.testql.toon.yaml
+- Update testql-scenarios/generated-from-pytests.testql.toon.yaml
+- Update tests/test_cli.py
+
+### Other
+- Update app.doql.less
+- Update planfile.yaml
+- Update prefact.yaml
+- Update project/analysis.toon.yaml
+- Update project/calls.mmd
+- Update project/calls.png
+- Update project/calls.toon.yaml
+- Update project/calls.yaml
+- Update project/compact_flow.mmd
+- Update project/compact_flow.png
+- ... and 11 more files
 
 ## [0.1.1] - 2026-06-17